+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 [Unreleased]
 
+## Added
+
+- Media endpoints
+    - `Mastodon.upload` for attaching media to a post
+    - `Mastodon.media` for getting information about a media attachment
+
 ---
 
-[0.0.1] - 2023-03-01
+[0.0.1] - 2023-03-03
 
 Initial release